Got all backend tests passing
We have 3 tests still failing on the front-end and a pa11y test failing
as well.

Tests are failing because of improper access and buttons that are hidden
from user unless logged in with editing capability.

Jira ticket: CAMS-406
