+++ backend/functions/lib/controllers/case-assignment/case.assignment.controller.test.ts
@@ -1,17 +1,30 @@
 import { CaseAssignmentController } from './case.assignment.controller';
 import { THROW_PERMISSIONS_ERROR_CASE_ID } from '../../testing/testing-constants';
-import { MockData } from '../../../../../common/src/cams/test-utilities/mock-data';
+import { MANHATTAN, MockData } from '../../../../../common/src/cams/test-utilities/mock-data';
 import { CaseAssignmentUseCase } from '../../use-cases/case.assignment';
 import { CamsError } from '../../common-errors/cams-error';
 import { ForbiddenError } from '../../common-errors/forbidden-error';
-import { createMockApplicationContext } from '../../testing/testing-utilities';
+import {
+  createMockApplicationContext,
+  createMockApplicationContextSession,
+} from '../../testing/testing-utilities';
+import { CamsRole } from '../../../../../common/src/cams/session';
 
 describe('Case Assignment Creation Tests', () => {
   const trialAttorneyRole = 'TrialAttorney';
   let applicationContext;
 
+  const user = {
+    name: 'Mock Name',
+    offices: [MANHATTAN],
+    roles: [CamsRole.CaseAssignmentManager],
+  };
   beforeEach(async () => {
     applicationContext = await createMockApplicationContext();
+    applicationContext.session = await createMockApplicationContextSession({ user });
+    // jest
+    //   .spyOn(ContextCreator, 'getApplicationContextSession')
+    //   .mockResolvedValue(MockData.getCamsSession());
   });
 
   afterEach(() => {
@@ -21,7 +34,7 @@ describe('Case Assignment Creation Tests', () => {
   test('A case is assigned to an attorney when requested', async () => {
     const listOfAttorneyNames = ['Jane'];
     const testCaseAssignment = {
-      caseId: '001-18-12345',
+      caseId: '081-18-12345',
       listOfAttorneyNames,
       role: trialAttorneyRole,
     };
@@ -44,7 +57,7 @@ describe('Case Assignment Creation Tests', () => {
   test('should assign all attorneys in the list', async () => {
     const listOfAttorneyNames = ['Jane', 'Tom', 'Adrian'];
     const testCaseAssignment = {
-      caseId: '001-18-12345',
+      caseId: '081-18-12345',
       listOfAttorneyNames,
       role: trialAttorneyRole,
     };
@@ -67,7 +80,7 @@ describe('Case Assignment Creation Tests', () => {
   test('should create only one assignment per attorney', async () => {
     const listOfAttorneyNames = ['Jane', 'Tom', 'Jane', 'Adrian', 'Tom'];
     const testCaseAssignment = {
-      caseId: '001-18-12345',
+      caseId: '081-18-12345',
       listOfAttorneyNames,
       role: trialAttorneyRole,
     };
@@ -99,7 +112,7 @@ describe('Case Assignment Creation Tests', () => {
       .mockResolvedValue(assignments);
 
     const assignmentController = new CaseAssignmentController(applicationContext);
-    const result = await assignmentController.getTrialAttorneyAssignments('001-18-12345');
+    const result = await assignmentController.getTrialAttorneyAssignments('081-18-12345');
     expect(result).toEqual(assignmentResponse);
   });
 
@@ -110,7 +123,7 @@ describe('Case Assignment Creation Tests', () => {
       .mockRejectedValue(new CamsError('TEST', { message: errorMessage }));
     const assignmentController = new CaseAssignmentController(applicationContext);
 
-    await expect(assignmentController.getTrialAttorneyAssignments('001-18-12345')).rejects.toThrow(
+    await expect(assignmentController.getTrialAttorneyAssignments('081-18-12345')).rejects.toThrow(
       errorMessage,
     );
   });
@@ -121,15 +134,23 @@ describe('Case Assignment Creation Tests', () => {
       listOfAttorneyNames: [],
       role: trialAttorneyRole,
     };
+    const rejectedAssignmentResponse = {
+      success: false,
+      message: 'User does not have appropriate access to create assignments.',
+      count: 0,
+      body: [],
+    };
+    const mockContext = await createMockApplicationContext();
+    mockContext.session = await createMockApplicationContextSession();
+
     jest
       .spyOn(CaseAssignmentUseCase.prototype, 'findAssignmentsByCaseId')
       .mockRejectedValue(new ForbiddenError('TEST_MODULE', { message: 'forbidden' }));
 
-    const assignmentController = new CaseAssignmentController(applicationContext);
-
-    await expect(
-      assignmentController.createTrialAttorneyAssignments(testCaseAssignment),
-    ).rejects.toThrow('Failed to authenticate to Azure');
+    const assignmentController = new CaseAssignmentController(mockContext);
+    const assignmentResponse =
+      await assignmentController.createTrialAttorneyAssignments(testCaseAssignment);
+    expect(assignmentResponse).toEqual(rejectedAssignmentResponse);
   });
 
   test('should throw any other errors on findAssignmentsByCaseId', async () => {
@@ -138,7 +159,7 @@ describe('Case Assignment Creation Tests', () => {
       .mockRejectedValue(new Error('An error'));
     const assignmentController = new CaseAssignmentController(applicationContext);
 
-    await expect(assignmentController.getTrialAttorneyAssignments('001-18-12345')).rejects.toThrow(
+    await expect(assignmentController.getTrialAttorneyAssignments('081-18-12345')).rejects.toThrow(
       'Unknown error',
     );
   });
